Enable SymlinkingUnitTestCase on Linux

+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/deployment/structure/war/symbolicLink/SymlinkingUnitTestCase.java
@@ -6,6 +6,7 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 
 import org.junit.Assert;
+import org.junit.Assume;
 import org.apache.commons.lang.SystemUtils;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
@@ -19,7 +20,6 @@ import org.jboss.logging.Logger;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -33,7 +33,6 @@ import org.junit.runner.RunWith;
 
 @RunWith(Arquillian.class)
 @RunAsClient
-@Ignore("AS7-3414")
 public class SymlinkingUnitTestCase {
 
     private static final Logger logger = Logger.getLogger(SymlinkingUnitTestCase.class);
@@ -46,6 +45,7 @@ public class SymlinkingUnitTestCase {
 
     @BeforeClass
     public static void beforeClass() throws IOException, InterruptedException {
+        Assume.assumeFalse(SystemUtils.IS_OS_WINDOWS);
         logger.infof("beforeClass() call.");
         // We are going to check whether or not the war deployment actually exists.
         Assert.assertTrue(checkForDeployment());
@@ -55,11 +55,15 @@ public class SymlinkingUnitTestCase {
 
     @AfterClass
     public static void afterClass() throws IOException {
-        logger.infof("afterClass() call.");
-        // Delete the symlink
-        symbolic.delete();
-        controllerClient.close();
-        controllerClient = null;
+        if (SystemUtils.IS_OS_WINDOWS) {
+         // skip windows;
+        } else {
+            logger.infof("afterClass() call.");
+            // Delete the symlink
+            symbolic.delete();
+            controllerClient.close();
+            controllerClient = null;
+        }
     }
 
     @After
@@ -98,15 +102,11 @@ public class SymlinkingUnitTestCase {
     }
 
     @Test
-    public void testEnabled() {
+    public void testEnabled() throws IOException {
         //logger.infof("Testing enabled bit");
         // By default we should not be able to browse to the symlinked page.
         Assert.assertTrue(getURLcode("symbolic") == 404);
-        try {
-            setup(true);
-        } catch (IOException e) {
-            logger.fatalf("IOException caught while setting up in testEnabled()");
-        }
+        setup(true);
         // First make sure that we can browse to index.html. This should work in the enabled or disabled version.
         Assert.assertTrue(getURLcode("index") == 200);
         // Now with symbolic.html.
@@ -114,15 +114,11 @@ public class SymlinkingUnitTestCase {
     }
 
     @Test
-    public void testDisabled() {
+    public void testDisabled() throws IOException {
         logger.infof("Testing disabled bit.");
         // By default we should not be able to browse to the symlinked page.
         Assert.assertTrue(getURLcode("symbolic") == 404);
-        try {
-            setup(false);
-        } catch (IOException e) {
-            logger.fatalf("IOException caught while setting up in testDisabled()");
-        }
+        setup(false);
         // First make sure that we can browse to index.html. This should work in the enabled or disabled version.
         Assert.assertTrue(getURLcode("index") == 200);
         // Now with symbolic.html.
